async redux repeat

+++ redux/src/redux/operations.js
@@ -1,12 +1,17 @@
 import axios from "axios";
 import { addNotes, deleteNote } from "./actions";
 
-export const putTask = task => async () => {
-  const data = await axios.post(
-    "https://authorization-3e689.firebaseio.com/tasks.json",
-    task
-  );
-  console.log(data);
+export const putTask = task => async (dispath, getState) => {
+  if (getState().token) {
+    console.log("getState inner putTask", getState());
+    const data = await axios.post(
+      "https://authorization-3e689.firebaseio.com/tasks.json",
+      task
+    );
+    console.log(data);
+  } else {
+    console.log("you cant put new note");
+  }
 };
 
 export const getTask = () => async dispatch => {
@@ -15,6 +20,11 @@ export const getTask = () => async dispatch => {
   );
 
   console.log("all tasks", data);
+
+  if (!data.data) {
+    return dispatch(addNotes([]));
+  }
+
   const transform = Object.keys(data.data)
     .map(key => ({
       ...data.data[key],
@@ -22,13 +32,19 @@ export const getTask = () => async dispatch => {
     }))
     .reverse();
   console.log("transform", transform);
+
   dispatch(addNotes(transform));
 };
 
-export const delTask = id => async dispatch => {
-  const del = await axios.delete(
-    `https://authorization-3e689.firebaseio.com/tasks/${id}.json`
-  );
-  console.log("del", del);
-  await dispatch(deleteNote(id));
+export const delTask = id => async (dispatch, getState) => {
+  console.log("-- ! --", getState());
+  if (getState().token) {
+    const del = await axios.delete(
+      `https://authorization-3e689.firebaseio.com/tasks/${id}.json`
+    );
+    console.log("del", del);
+    await dispatch(deleteNote(id));
+  } else {
+    console.log("you cant delete");
+  }
 };
